Papildoma optimizacija ir taimerio pataisymas
